recursos/leccion2/postgresql/sistema/requirements.txt
# Install the packages from PyPi repository.
python-dotenv==1.0.1
psycopg2==2.9.5
